README
This game is played on 1 blink.

Double click to start the game.

You as a player is only allowed to press when the blink goes green. When the blink goes blue you have to wait until the blink turns back to green. 

If you fail to click on the green blink too late or if you clink on the blue blink you will fail the game.