README
For static, I wanted to make a emotion of calm and hope. 
To express that emotion, I decided to use the spectrum of Green color that represent calmness and hope. 
However, the blinks were not as sensitive to colors as expected, which didn't create a straightforward spectrum, but the color difference is still present.