Example of GameStates shared across entire cluster
This code abstracts the signalState and the game mode loops. Listens for updates in game mode, and the game mode code simply runs in its individual loops. Changing game modes is based on a single click here, but can be based on time, or any other logic. Using the function changeMode() and passing it the mode you want to go to will have the whole cluster change together.