Adding an example of a timed game state across all Blinks
This expands on the shared game state to show how gameState can be changed by any kind of user input or game logic. Here, the game changes state at the end of a game timer, but requires button presses to start the game and return to ready mode from an end game state.