Github Tutorial Upstream Update
Add a PDF to explain how to keep your repo up to date with the class repo. This is really handy because all of the examples, and class work is updated on your machine to be the latest. Of course, continue to use pull requests to submit assignments. Do this every week or as often as you'd like so that your materials are up to date.