Update id.ino
First cleanup. Take the click listener out of the for loop. The click only needs to be listened for once, not for each face.

Then, added a variable for our own Id index, since each Blink needs to know which of the ids is theirs.

Created an unassigned value (could have used a boolean to determine if we have received an id yet also, but checking to see if we still == UNASSIGNED feels more straightforward)

Only look to neighbors for assignment if I am in fact, UNASSIGNED

Then the display. No need to loop through the array, since we simply want to display our own id, which we can access from the index we have been assigned.

Add a display state for when we are unassigned as well, for fun, let's us know the Blinks are on and working.